Backend/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
colorlog==6.7.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.1.0
django-postgresql==0.0.3
django-request==1.6.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
idna==3.4
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4